refactor(hooks): split host hook integrations
