+++ tests/agent_suite/agent_test.rs
@@ -621,14 +621,15 @@ fn valid_single_quoted_yaml_scalar(value: &str) -> bool {
 
 #[test]
 fn generated_prompt_rules_do_not_hardcode_repo_local_graph_db() {
+    // Hosts that render their own rule text, plus the shared renderer that
+    // copilot/gemini/opencode/kimi/vibe delegate to.
     for (name, source) in [
         ("claude", include_str!("../../src/agents/claude.rs")),
-        ("copilot", include_str!("../../src/agents/copilot.rs")),
-        ("gemini", include_str!("../../src/agents/gemini.rs")),
         ("kiro", include_str!("../../src/agents/kiro.rs")),
-        ("kimi", include_str!("../../src/agents/kimi.rs")),
-        ("opencode", include_str!("../../src/agents/opencode.rs")),
-        ("vibe", include_str!("../../src/agents/vibe.rs")),
+        (
+            "prompt_rules",
+            include_str!("../../src/agents/prompt_rules.rs"),
+        ),
     ] {
         assert!(
             !source.contains(".tracedecay/tracedecay.db"),
@@ -638,6 +639,22 @@ fn generated_prompt_rules_do_not_hardcode_repo_local_graph_db() {
             source.contains("tracedecay_active_project")
                 && source.contains("tracedecay_storage_status"),
             "{name} generated guidance should point store questions to active-project/storage-status tools"
+        );
+    }
+    for (name, source) in [
+        ("copilot", include_str!("../../src/agents/copilot.rs")),
+        ("gemini", include_str!("../../src/agents/gemini.rs")),
+        ("kimi", include_str!("../../src/agents/kimi.rs")),
+        ("opencode", include_str!("../../src/agents/opencode.rs")),
+        ("vibe", include_str!("../../src/agents/vibe.rs")),
+    ] {
+        assert!(
+            !source.contains(".tracedecay/tracedecay.db"),
+            "{name} generated guidance must not hardcode the repo-local graph DB path"
+        );
+        assert!(
+            source.contains("standard_prompt_rules"),
+            "{name} should delegate prompt rules to the shared renderer"
         );
     }
 }
@@ -1040,6 +1057,16 @@ fn test_hermes_local_install_writes_profile_plugin() {
 
     let skill = std::fs::read_to_string(plugin_dir.join("skills/tracedecay/SKILL.md")).unwrap();
     assert!(skill.contains("Use tracedecay"));
+    // CLI-fallback steering: mirrors CLI_FALLBACK_PROMPT_RULES, worded for the
+    // Hermes plugin surface (tool calls already shell out to the tracedecay CLI).
+    assert!(
+        skill.contains("tracedecay tool ") && skill.contains("--help"),
+        "Hermes skill must steer the agent to the `tracedecay tool` CLI when a tool invocation fails"
+    );
+    assert!(
+        skill.contains("instead of querying"),
+        "Hermes skill must warn against querying .tracedecay databases directly"
+    );
 
     assert_hermes_config_enables_tracedecay_memory(&project.path().join(".hermes/config.yaml"));
     assert!(
